src/ai/face_analyzer.py
import requests
import numpy as np
from pathlib import Path
from typing import List, Dict, Any
from PIL import Image
import io
import random
from utils.logger import setup_logger

logger = setup_logger()

class FaceAnalyzer:
    def __init__(self):
        self.known_faces = {}
    
    def analyze_profile_image(self, image_url: str, username: str) -> Dict[str, Any]:
        """Analyze profile image - demo implementation"""
        try:
            logger.info(f"Analyzing profile image for {username} (demo mode)")
            
            # Simulate face detection
            has_face = random.choice([True, False])
            faces_detected = random.randint(0, 2) if has_face else 0
            
            # Store demo face data
            if has_face:
                self.known_faces[username] = f"demo_face_encoding_{username}"
            
            return {
                'faces_detected': faces_detected,
                'face_locations': [(50, 150, 200, 100)] if has_face else [],
                'has_face': has_face,
                'image_dimensions': (400, 400),
                'demo_mode': True
            }
            
        except Exception as e:
            logger.error(f"Face analysis error: {e}")
            return {'faces_detected': 0, 'has_face': False, 'error': str(e)}
    
    def compare_faces(self, username1: str, username2: str) -> Dict[str, Any]:
        """Compare faces between two profiles - demo implementation"""
        if username1 not in self.known_faces or username2 not in self.known_faces:
            return {'match': False, 'confidence': 0.0, 'error': 'Face data not available'}
        
        try:
            # Simulate face comparison
            confidence = random.uniform(0.3, 0.9)
            match = confidence > 0.7
            
            return {
                'match': match,
                'confidence': confidence,
                'distance': 1 - confidence,
                'demo_mode': True
            }
            
        except Exception as e:
            logger.error(f"Face comparison error: {e}")
            return {'match': False, 'confidence': 0.0, 'error': str(e)}